+++ app/index.html
@@ -21,10 +21,12 @@
 				</form>
 			</div>
 			<div class="pure-u-5-5">
-				<p class="lucky js-random-article">I'm feeling lucky</p>
+				<p class="lucky">
+					<a href="https://en.wikipedia.org/wiki/Special:Random" target="_blank">I'm feeling lucky</a>
+				</p>
 			</div>
 		</div>
-		<div class="pure-g article-container"></div>
+		<div class="pure-g article-container js-article-container"></div>
 	</div>
 </body>
 </html>
