+++ app/index.js
@@ -1,56 +1,28 @@
 
-/* eslint-disable no-console */
-import { Observable } from 'rx';
-import fetchJsonp from 'fetch-jsonp';
+import 'rx-dom';
+import Rx from 'rx';
 
-const URL = 'http://en.wikipedia.org/w/api.php?' +
-			'format=json&' +
-			'action=query&' +
-			'generator=search&' +
-			'gsrnamespace=0&' +
-			'gsrlimit=10&' +
-			'prop=pageimages|extracts&' +
-			'pilimit=max&' +
-			'exintro&' +
-			'explaintext&' +
-			'exsentences=1&' +
-			'exlimit=max&' +
-			'callback=?&' +
-			'gsrsearch=javascript';
+import {
+	getURL,
+	createArticle,
+	makeCall
+} from 'helpers';
 
-const requestStream = Observable.just(URL);
-const responseStream = requestStream
-						.flatMap(() => Observable.fromPromise(fetchJsonp(URL)))
-						.flatMap(response => response.json())
-						.map(response => response.query.pages);
+const input = document.querySelector('.js-search');
+const throttledInput$ = Rx.DOM
+						.keyup(input)
+						.pluck('target', 'value')
+						.filter((text) => text.length > 3)
+						.debounce(500)
+						.distinctUntilChanged();
 
+const searchStream$ = throttledInput$.flatMapLatest(userSearch => makeCall(getURL(userSearch)));
 
-function createArticle(searchResult) {
-	const {
-		title,
-		extract
-	} = searchResult;
-
-	const article = `
-		<header>
-			<h2>${title}</h2>
-			<p>${extract}</p>
-			<a href="http://en.wikipedia.org/wiki/${title.replace(' ', '_')}" target="_blank">See it on wikipedia</a>
-		</header>
-	`;
-
-	const element = document.createElement('article');
-	element.classList = 'pure-u-5-5 article';
-	element.innerHTML = article;
-
-	return element;
-}
-
-responseStream.subscribe((articles) => {
-	const container = document.querySelector('.article-container');
+searchStream$.subscribe((articles) => {
+	const container = document.querySelector('.js-article-container');
+	container.innerHTML = '';
 
 	Object.keys(articles).forEach((article) => {
 		container.appendChild(createArticle(articles[article]));
 	});
 });
-/* eslint-enable no-console */
